requirements.txt
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-http>=1.20.0
requests>=2.31.0
anthropic>=0.39.0
clickhouse-connect>=0.7.0
python-dotenv>=1.0.0
psutil>=5.9.0
